Fix #3 (see last commit)

+++ db.js
@@ -230,9 +230,9 @@ async function getTransactionsForFriendsFeed(friendIDs, myID, before = null, aft
     WHERE (actor_id = ANY ($1::integer[]) OR target_id = ANY ($1::integer[]))
       AND (audience = 'friends' OR audience = 'public' OR actor_id = $2 OR target_id = $2)
       AND status = 'settled'
-      AND ($6 IS NULL OR id < $6)
-      AND ($3 IS NULL OR date_completed < $3)
-      AND ($4 IS NULL OR date_completed > $4)
+      AND ($6::integer IS NULL OR id < $6)
+      AND ($3::integer IS NULL OR date_completed < $3)
+      AND ($4::integer IS NULL OR date_completed > $4)
     ORDER BY id DESC NULLS LAST
     LIMIT $5
     `, [friendIDs, myID, before, after, limit, lastTransactionID]);
